+++ src/components/pages/Home.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
-import Main from './main';
-import Footer from './footer';
-import {Intro} from './Intro';
+import Main from '../Main';
+import Footer from '../Footer';
+import {Intro} from '../Intro';
 
 class Home extends Component {
   render() {
